% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{aggregate_day}
\alias{aggregate_day}
\title{Aggregate dataframe by day (i.e. all trajectories on Monday)}
\usage{
aggregate_day(traj, day)
}
\arguments{
\item{traj}{trajectory df}

\item{day}{day to aggregate by (1 = Sunday, 2 = Monday, ... , 7 = Saturday)}
}
\value{
aggregated data frame
}
\description{
Aggregate dataframe by day (i.e. all trajectories on Monday)
}
