% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{aggregate_raster_region}
\alias{aggregate_raster_region}
\title{Aggregate raster to region of interest}
\usage{
aggregate_raster_region(
  raster,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{xmin}{min x value}

\item{xmax}{max x calue}

\item{ymin}{min y value}

\item{ymax}{max y value}

\item{rasterized}{rasterized object}
}
\value{
Cropped raster to ROI
}
\description{
Aggregate raster to region of interest
}
