% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{aggregate_sf_roi}
\alias{aggregate_sf_roi}
\title{Aggregate sf data frame to region of interest}
\usage{
aggregate_sf_roi(df, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL)
}
\arguments{
\item{df}{sf data frame of trajectories with geometry column or sfTrack or sfTracks}

\item{xmin}{min x}

\item{xmax}{max x}

\item{ymin}{min y}

\item{ymax}{max y}
}
\value{
Aggregated data frame
}
\description{
Aggregate sf data frame to region of interest
}
