% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{aggregate_sft_time}
\alias{aggregate_sft_time}
\title{Aggregate sfTrack by time}
\usage{
aggregate_sft_time(sftrack, from, to)
}
\arguments{
\item{sftrack}{sfTrack}

\item{from}{from in posixct format}

\item{to}{to in posxict format}
}
\value{
Returns aggregated sfTrack
}
\description{
Aggregate sfTrack by time
}
