% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{animate_single_track}
\alias{animate_single_track}
\title{Animate single trajectory using movevis (BEWARE OF MEMORY/RENDERING PROBLEMS)}
\usage{
animate_single_track(
  trajectory,
  res,
  filename = "trajectory.gif",
  unit = "min"
)
}
\arguments{
\item{trajectory}{singular trajectory data frame}

\item{res}{temporal resolution (i.e. 5 = 5 mins)}

\item{filename}{filename for output GIF}

\item{unit}{units for temporal resolution (minutes at default)}
}
\value{
animation of trajectory in GIF
}
\description{
Animate single trajectory using movevis (BEWARE OF MEMORY/RENDERING PROBLEMS)
}
