% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{as.sf.Tracks}
\alias{as.sf.Tracks}
\title{Convert sf data to Track}
\usage{
as.sf.Tracks(df)
}
\arguments{
\item{df}{Trajectory data frame in sf and sftime format to be converted to Track}
}
\value{
Track
}
\description{
Convert sf data to Track
}
