% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{calendar_heatmap}
\alias{calendar_heatmap}
\title{Calendar heatmap of trajectories data frame}
\usage{
calendar_heatmap(df, value)
}
\arguments{
\item{df}{data frame of trajectories}

\item{value}{value for heatmap}
}
\value{
calendar heatmap
}
\description{
Calendar heatmap of trajectories data frame
}
