% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{density_heatmap}
\alias{density_heatmap}
\title{Plot kernel density heat map of trajectory measurements}
\usage{
density_heatmap(df, value, resolution, date)
}
\arguments{
\item{df}{trajectories data frame or sfTrack or sfTracks}

\item{value}{value desired to make heat map off}

\item{resolution}{desired resolution}

\item{date}{optional parameter to create a day heatmap with 6 plots and 4 hour intervals}
}
\value{
plot of density heat map
}
\description{
Plot kernel density heat map of trajectory measurements
}
