% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{df_to_sfTracks}
\alias{df_to_sfTracks}
\title{Coerce nested sf data frame to sfTracks}
\usage{
df_to_sfTracks(df)
}
\arguments{
\item{df}{data frame to coerce}
}
\value{
sfTracks object
}
\description{
Coerce nested sf data frame to sfTracks
}
