% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{find_intersections_density}
\alias{find_intersections_density}
\title{Find density of intersections by returning raster layer of intersections}
\usage{
find_intersections_density(df, resolution)
}
\arguments{
\item{df}{sf data frame or sfTrack or sfTracks}

\item{resolution}{desired resolution}
}
\value{
rasterized intersections
}
\description{
Find density of intersections by returning raster layer of intersections
}
