% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{geodata_to_sf}
\alias{geodata_to_sf}
\title{Convert trajectory data frame in  or lat long format to sf}
\usage{
geodata_to_sf(df, identifier)
}
\arguments{
\item{df}{A trajectory data frame with a geometry column or in lat long format}

\item{identifier}{Unique identifier to group data frame by}
}
\value{
A nested data frame in sf format
}
\description{
Convert trajectory data frame in  or lat long format to sf
}
