% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{gi_hotspot}
\alias{gi_hotspot}
\title{Getis-ord hotspot analysis}
\usage{
gi_hotspot(poly_points)
}
\arguments{
\item{poly_points}{polygon points from ppa_polygons function}
}
\value{
heatmap
}
\description{
Getis-ord hotspot analysis
}
