% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{idwi_raster}
\alias{idwi_raster}
\title{Interpolate raster using inverse distance weighted interpolation}
\usage{
idwi_raster(df, measurement, resolution)
}
\arguments{
\item{df}{data frame}

\item{measurement}{measurement to rasterize off}

\item{resolution}{desired resolution}
}
\value{
Interpolated raster layer
}
\description{
Interpolate raster using inverse distance weighted interpolation
}
