% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{plot_day}
\alias{plot_day}
\title{Plot values by weekday}
\usage{
plot_day(df, value, xmin, xmax, ymin, ymax)
}
\arguments{
\item{df}{trajectories data frame}

\item{value}{optional parameter to show measurement value}

\item{xmin}{min x}

\item{xmax}{max x}

\item{ymin}{min y}

\item{ymax}{max y}
}
\value{
plot of aggregated values
}
\description{
Plot values by weekday
}
