% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{plot_hour}
\alias{plot_hour}
\title{Plot values by hour}
\usage{
plot_hour(df, value, xmin, xmax, ymin, ymax)
}
\arguments{
\item{df}{trajectories data frame}

\item{xmin}{min x}

\item{xmax}{max x}

\item{ymin}{min y}

\item{ymax}{max y}
}
\value{
plot of aggregated values
}
\description{
Plot values by hour
}
