% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{plot_hour_density}
\alias{plot_hour_density}
\title{Visualize density in region of interest by time of day}
\usage{
plot_hour_density(df, xmin, xmax, ymin, ymax)
}
\arguments{
\item{df}{trajectories data frame}

\item{xmin}{min x}

\item{xmax}{max x}

\item{ymin}{min y}

\item{ymax}{max y}
}
\value{
plot of aggregated values
}
\description{
Visualize density in region of interest by time of day
}
