% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot trajectories using ggplot2}
\usage{
plot_traj(trajectories, value)
}
\arguments{
\item{trajectories}{trajectories data frame}

\item{value}{value to base scale off}
}
\value{
ggplot2 of trajectories
}
\description{
Plot trajectories using ggplot2
}
