% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotExpression}
\alias{plotExpression}
\title{Plot gene expression along pseudotime.}
\usage{
plotExpression(counts, sds, gene, ...)
}
\arguments{
\item{...}{parameters including:}
}
\description{
Plot gene expression along pseudotime.
}
