% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{ppa_polygons}
\alias{ppa_polygons}
\title{Polygonal raster pattern analysis}
\usage{
ppa_polygons(df, value, res)
}
\arguments{
\item{df}{trajectories data frame in sf format}

\item{value}{value to make raster poly}

\item{res}{resolution}
}
\value{
returns polygon dataframe
}
\description{
Polygonal raster pattern analysis
}
