% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{raster_track}
\alias{raster_track}
\title{Rasterize track data}
\usage{
raster_track(track, value, resolution)
}
\arguments{
\item{track}{track}

\item{value}{value wanted to raster with}

\item{resolution}{desired resolution}
}
\value{
rasterized object
}
\description{
Rasterize track data
}
