% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sf_to_raster_stars}
\alias{sf_to_raster_stars}
\title{sf to stars raster}
\usage{
sf_to_raster_stars(df, value)
}
\arguments{
\item{df}{Trajectory data frame in sf format}

\item{value}{Data measurements to rasterize}
}
\value{
stars object
}
\description{
sf to stars raster
}
