% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sf_to_rasterize}
\alias{sf_to_rasterize}
\title{sf trajectory data frame to raster with selected properties to rasterize}
\usage{
sf_to_rasterize(df, data, resolution, from, to)
}
\arguments{
\item{df}{Trajectory data frame in sf format to rasterize}

\item{data}{Data values wanted to rasterize}

\item{resolution}{Level of resolution}

\item{from}{Optional parameter from in as.POSIXct format to aggregate data from}

\item{to}{Optional parameter to in as.POSIXct format to aggregate to}
}
\value{
rasterized object
}
\description{
sf trajectory data frame to raster with selected properties to rasterize
}
