% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Get XY coordinates from sf object (taken from jmlondon at https://github.com/r-spatial/sf/issues/231)}
\usage{
sfc_as_cols(x, names = c("x", "y"))
}
\arguments{
\item{x}{data frame in sf format}

\item{names}{names to name XY columns}
}
\value{
data frame with XY columns
}
\description{
Get XY coordinates from sf object (taken from jmlondon at https://github.com/r-spatial/sf/issues/231)
}
