% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sfcube}
\alias{sfcube}
\title{Plot space time cube of sf trajectory}
\usage{
sfcube(df)
}
\arguments{
\item{df}{data frame sf trajectory}
}
\value{
space time cube
}
\description{
Plot space time cube of sf trajectory
}
