% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sft.plot_ts}
\alias{sft.plot_ts}
\title{Plot time series of track}
\usage{
sft.plot_ts(sft, value)
}
\arguments{
\item{sft}{sfTrack}

\item{value}{y value}
}
\value{
ggplot time series
}
\description{
Plot time series of track
}
