% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-sfTrack.R
\docType{class}
\name{sfTrack-class}
\alias{sfTrack-class}
\title{A Track in sf format}
\description{
A Track in sf format
}
\section{Slots}{

\describe{
\item{\code{id}}{unique id value for track}

\item{\code{data}}{dataframe of track data}

\item{\code{time}}{POSIXct format timestamps}

\item{\code{geometry}}{geometry of track in sfc format}

\item{\code{line}}{LINESTRING trajectory of track}
}}

