% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{sfts.plot_violin}
\alias{sfts.plot_violin}
\title{Plot violin plot of sfTracks value}
\usage{
sfts.plot_violin(sfts, value)
}
\arguments{
\item{sfts}{sfTracks}

\item{value}{y calue}
}
\value{
violin plot
}
\description{
Plot violin plot of sfTracks value
}
