% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{traj_heatmap}
\alias{traj_heatmap}
\title{Plot kernel density heat map of trajectories}
\usage{
traj_heatmap(trajectories)
}
\arguments{
\item{trajectories}{trajectories dataframe}
}
\value{
kernel density heatmap
}
\description{
Plot kernel density heat map of trajectories
}
