% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{traj_quadrat}
\alias{traj_quadrat}
\title{Plot quadrat intensity of points of a trajectory}
\usage{
traj_quadrat(df)
}
\arguments{
\item{df}{sf trajectories data frame}
}
\value{
plot of density heat map
}
\description{
Plot quadrat intensity of points of a trajectory
}
