% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{cluster_traj}
\alias{cluster_traj}
\title{Cluster trajectories}
\usage{
cluster_traj(trajectories, num_clusters)
}
\arguments{
\item{trajectories}{trajectories data frame in sf format}

\item{num_clusters}{desired number of clusters}
}
\value{
Returns clustered trajectories data frame
}
\description{
Cluster trajectories
}
