% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traviz-methods.R
\name{density_heatmap}
\alias{density_heatmap}
\title{Plot kernel density heat map of trajectory measurements}
\usage{
density_heatmap(trajectories, value, resolution)
}
\arguments{
\item{trajectories}{trajectories data frame}

\item{value}{value desired to make heat map off}

\item{resolution}{desired resolution}
}
\value{
plot of density heat map
}
\description{
Plot kernel density heat map of trajectory measurements
}
