% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotExpression.R
\name{plotExpression}
\alias{plotExpression}
\alias{plotExpression,matrix,SlingshotDataSet,character-method}
\alias{plotExpression,matrix,PseudotimeOrdering,character-method}
\title{Plot gene expression along pseudotime.}
\usage{
plotExpression(counts, sds, gene, ...)

\S4method{plotExpression}{matrix,SlingshotDataSet,character}(
  counts,
  sds,
  gene,
  type = "loess",
  span = 0.75,
  alpha = 1,
  lwd = 1,
  size = 2/3
)

\S4method{plotExpression}{matrix,PseudotimeOrdering,character}(
  counts,
  sds,
  gene,
  type = "loess",
  span = 0.75,
  alpha = 1,
  lwd = 1,
  size = 2/3
)
}
\arguments{
\item{counts}{The matrix of gene expression counts.}

\item{sds}{A \code{SlingshotDataSet} or \code{PseudotimeOrdering} object,
typically obtained after trajectory inference using \code{Slingshot}.}

\item{gene}{Gene name of gene to plot.}

\item{...}{parameters including:}

\item{type}{The type of smoother. Defaults to \code{"loess"}.}

\item{span}{If \code{type} is \code{"loess}, the \code{span} of the smoother.
See \code{loess} documentation.}

\item{alpha}{Numeric between 0 and 1, determines the transparency of data points,
see \code{scale_color_viridis_d}.}

\item{lwd}{Line width of the smoother. Passed to \code{\link{geom_line}}.}

\item{size}{Character expansion of the data points. Passed to \code{\link{geom_point}}.}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot gene expression along pseudotime.

Plots a fast loess smoother of gene expression for each lineage.
}
\examples{
library(ggplot2)
data(crv, package="traffic")
data(counts, package="traffic")
plotExpression(counts = counts, sds=crv, gene=rownames(counts)[1])
}
